Fix Win64 build
R=verwaest@chromium.org

Review URL: https://codereview.chromium.org/12221012

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@13594 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/deoptimizer-x64.cc
@@ -539,7 +539,7 @@ void Deoptimizer::DoCompiledStubFrame(TranslationIterator* iterator,
       Smi::FromInt(StackFrame::STUB_FAILURE_TRAMPOLINE));
   output_frame->SetFrameSlot(output_frame_offset, value);
 
-  int caller_arg_count = 0;
+  intptr_t caller_arg_count = 0;
   if (descriptor->stack_parameter_count_ != NULL) {
     caller_arg_count =
         input_->GetRegister(descriptor->stack_parameter_count_->code());
